+++ Sources/DiagnosticsLogger.swift
@@ -143,6 +143,9 @@ extension DiagnosticsLogger {
                 return assertionFailure()
         }
 
+        // Make sure we have more then 500MB space left
+        guard UIDevice.current.freeDiskSpaceInBytes > (500 * 1024) else { return }
+
         fileHandle.seekToEndOfFile()
         fileHandle.write(data)
         logSize += Int64(data.count)