+++ Sources/DiagnosticsLogger.swift
@@ -30,7 +30,7 @@ public final class DiagnosticsLogger {
 
     private lazy var formatter: DateFormatter = {
         let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd hh:mm:ss"
+        formatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
         formatter.locale = Locale(identifier: "en_US")
         formatter.timeZone = TimeZone(identifier: "GMT")!
         return formatter