+++ Changelog.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 ### Next
+- Changed log times to use 24H clock (#40)
 - Support for iOS 10.0. #42
 
 ### 1.2.0