+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ### Next
 - Added public static method  `DiagnosticsLogger.isSetUp()` (#41)
+- Changed log times to use 24H clock (#40)
 - Support for iOS 10.0. #42
 
 ### 1.2.0