+++ Changelog.md
@@ -1,7 +1,9 @@
 ## Changelog
 
 ### Next
-- Support for iOS 10.0 and update package.swift & Diagnostics.podspec to minimum deployment target iOS 10.0. #42
+- Added public static method  `DiagnosticsLogger.isSetUp()` (#41)
+- Changed log times to use 24H clock (#40)
+- Support for iOS 10.0 and update package.swift & Diagnostics.podspec to minimum deployment target iOS 10.0. #42 #47
 
 ### 1.2.0
 - Forward the output of the DiagnosticsLogger to not disable default logging