+++ Changelog.md
@@ -3,7 +3,7 @@
 ### Next
 - Added public static method  `DiagnosticsLogger.isSetUp()` (#41)
 - Changed log times to use 24H clock (#40)
-- Support for iOS 10.0. #42
+- Support for iOS 10.0 and update package.swift & Diagnostics.podspec to minimum deployment target iOS 10.0. #42 #47
 
 ### 1.2.0
 - Forward the output of the DiagnosticsLogger to not disable default logging