+++ Changelog.md
@@ -1,6 +1,8 @@
 ## Changelog
 
 ### Next
+
+### 1.3.0
 - Make sure to check whether there's enough space left to save new logs (#37)
 - Added public static method  `DiagnosticsLogger.isSetUp()` (#41)
 - Changed log times to use 24H clock (#40)